requirements.txt
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
groq>=0.12.0
pyyaml>=6.0
python-dotenv>=1.0.0
aiosqlite>=0.20.0
telethon>=1.36.0
